requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorlog==4.1.0
ConfigArgParse==0.15.1
cryptography==2.8
enum34==1.1.6
Flask==1.1.1
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
locustio==0.13.5
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==8.1.0
msgpack-python==0.5.6
namedlist==1.7
packaging==20.0
paramiko==2.7.1
pluggy==0.13.1
py==1.8.1
pycparser==2.19
pymongo==3.10.1
PyNaCl==1.3.0
pyparsing==2.4.6
pytest==5.3.2
pyzmq==18.1.1
requests==2.22.0
six==1.13.0
sshtunnel==0.1.5
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==0.16.0
zipp==1.0.0
arrow==0.15.5
